requirements.txt
Flask
torch
torchvision
pillow
